seperate db connection from calls
