+++ lib/LoginServer/index.js
@@ -20,7 +20,7 @@ const packet = require('./packet');
 const { NPSUserStatus } = require('./npsUserStatus');
 
 const { Connection } = require("../../src/Connection");
-const database = require("../database");
+const database = require("../database/db");
 
 function npsGetCustomerIdByContextId(contextId) {
