+++ lib/database/index.js
@@ -16,7 +16,6 @@
 
 require('dotenv-safe').config();
 const { Pool } = require('pg')
-const { logger } = require("../../src/logger")
 
 // pools will use environment variables
 // for connection information
@@ -26,45 +25,5 @@ const pool = new Pool({
   user: process.env.DB_USER,
 })
 
-/**
- * Create the sessions database table if it does not exist
- * @param {Function} callback
- */
-async function createDB() {
 
-  return pool.connect().then(pool.query(
-    `CREATE TABLE IF NOT EXISTS sessions (customer_id INTEGER NOT NULL UNIQUE, 
-      session_key TEXT, s_key TEXT, context_id TEXT, connection_id INTEGER)`,
-    [],
-  )).catch((e) => {
-    if (e.code === "ENOTFOUND") {
-      logger.error(`Error connecting to database: ${e}`)
-      process.exit(1)
-    } else {
-      console.trace(e);
-    };
-  });
-}
-
-/**
- * Fetch session key from database based on remote address
- * @param {string} remoteAddress
- */
-async function fetchSessionKeyByConnectionId(connectionId) {
-  return pool.connect().then(
-    pool.query(
-      "SELECT session_key, s_key FROM sessions WHERE connection_id = $1",
-      [connectionId],
-    )).catch((e) => { console.trace(e); });
-}
-
-async function updateSessionKey(customerId, sessionKey, contextId, connectionId) {
-  const sKey = sessionKey.substr(0, 16);
-  return pool.connect().then(pool.query(
-    `INSERT INTO sessions (customer_id, session_key, s_key, context_id, 
-      connection_id) VALUES ($1, $2, $3, $4, $5) ON CONFLICT (customer_id) DO UPDATE SET session_key = $2, s_key = $3, context_id = $4, connection_id = $5`,
-    [customerId, sessionKey, sKey, contextId, connectionId],
-  )).catch((e) => { console.trace(e); });
-}
-
-module.exports = { createDB, updateSessionKey, updateSessionKey }
+module.exports = pool
