+++ src/TCPManager.js
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-const database = require('../lib/database');
+const database = require('../lib/database/db');
 const ClientConnectMsg = require('./ClientConnectMsg');
 const lobby = require('./lobby');
 const { logger } = require('./logger');
